requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
DateTime==4.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.0.3
zope.interface==5.4.0
